did correct logic
