+++ src/screens/ScreeStarShips/ScreenStarShips.tsx
@@ -1,7 +1,10 @@
-import React, {useState} from 'react';
+import React, {useState, useEffect} from 'react';
 import {View, Text} from 'react-native';
 import {useAppSelector} from '../../redux/hooks/hooks';
 import {RouteProp} from '@react-navigation/native';
+import {setStarships} from '../../redux/slices/starshipsCharastersSlice';
+import FetchStarShips from '../../utlis/FetchData/FetchStarShips';
+import {useAppDispatch} from '../../redux/hooks/hooks';
 
 type RootStackParamList = {
   ScreenStarShips: {name: string};
@@ -12,8 +15,16 @@ type ScreenStarShipsProps = {
 
 const ScreenStarShips = ({route}: ScreenStarShipsProps) => {
   const {name} = route.params;
+  const dispatch = useAppDispatch();
   const starShips = useAppSelector(state => state.starshipsData.starships);
-  const [isOpen, setIsOpen] = useState(false);
+
+  useEffect(() => {
+    const fetchStarShips = async () => {
+      const starShips = await FetchStarShips();
+      dispatch(setStarships(starShips));
+    };
+    fetchStarShips();
+  }, [dispatch]);
 
   return (
     <View>
