+++ src/screens/ScreenFilms/ScreenFilms.tsx
@@ -1,14 +1,12 @@
 import React, {useState, useEffect} from 'react';
 import {useAppSelector} from '../../redux/hooks/hooks';
 import {RouteProp} from '@react-navigation/native';
-import {
-  Container,
-  Block,
-  TextStyled,
-  StyledFlatList,
-} from './ScreenFilms.styles';
+import {Container} from './ScreenFilms.styles';
+import FetchFilms from '../../utlis/FetchData/FetchFilms';
+import {setFilms} from '../../redux/slices/filmsCharactersSlice';
 import FilmsDetails from '../../components/FilmsDetails/FilmsDetails';
 import {FilmsTypes} from '../../entites/types/FilmsTypes';
+import {useAppDispatch} from '../../redux/hooks/hooks';
 import {FlatList} from 'react-native';
 import useGetCharasterURL from '../../redux/hooks/customHooks';
 
@@ -21,8 +19,17 @@ type ScreenFilmsProps = {
 
 const ScreenFilms = ({route}: ScreenFilmsProps) => {
   const {name} = route.params;
+  const dispatch = useAppDispatch();
   const filmsData = useAppSelector(state => state.filmsData.films);
   const urlCharaster = useGetCharasterURL(name);
+
+  useEffect(() => {
+    const fetchFilms = async () => {
+      const films = await FetchFilms();
+      dispatch(setFilms(films));
+    };
+    fetchFilms();
+  }, [dispatch]);
 
   const result = filmsData.filter((item: FilmsTypes) => {
     return item.characters.some((url: string) => url === urlCharaster);
