+++ src/screens/ScreenVehicles/ScreenVehicles.tsx
@@ -1,7 +1,10 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import {View, Text} from 'react-native';
 import {useAppSelector} from '../../redux/hooks/hooks';
 import {RouteProp} from '@react-navigation/native';
+import {setVehicles} from '../../redux/slices/vehiclesCharastersSlice';
+import FetchVehicles from '../../utlis/FetchData/FetchVehicles';
+import {useAppDispatch} from '../../redux/hooks/hooks';
 
 type RootStackParamList = {
   ScreenVehicles: {name: string};
@@ -12,7 +15,16 @@ type ScreenStarShipsProps = {
 
 const ScreenVehicles = ({route}: ScreenStarShipsProps) => {
   const {name} = route.params;
+  const dispatch = useAppDispatch();
   const vehicles = useAppSelector(state => state.vehiclesData.vehicles);
+
+  useEffect(() => {
+    const fetchVehicles = async () => {
+      const vehicles = await FetchVehicles();
+      dispatch(setVehicles(vehicles));
+    };
+    fetchVehicles();
+  }, [dispatch]);
 
   return (
     <View>
